requirements.txt
python-docx
walkdir
pywin32
PyPDF2